Implement MouseSensorListener which uses the phone accelerometer to move the mouse
By using the inbuilt accelerometer in a phone, we can in theory make
the phone work like a traditional mouse. This is very difficult however
as the accelerometer sensor is not only noisy but also very inconsistent.
The current implementation does not work very well and needs further
development, although that might be unfeasible as it is unclear whether
the underlying acceleration data is even useable for this purpose.